Backend/delete-booking/requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==3.*
google-cloud-firestore